+++ src/js/main.js
@@ -3,8 +3,10 @@ let dirutils = require('./js/dir-utils.js');
 let openDirCommand = isWin ? 'explorer ' : 'open ';
 let cmdCommand = isWin ? "cmd /K cd /d "  : "terminal ";
 let childProcess = require("child_process");
-let autocompleteActions = require('./js/autocomplete-actions.js');
+let fieldActions = require('./js/field-actions.js');
 let onDirActions = require('./js/on-dir-actions.js');
+let autocompleteActions = require('./js/autocomplete-actions.js');
+
 
 dirutils.init(function(){
     $(function() {
@@ -27,38 +29,9 @@ dirutils.init(function(){
         }).blur(function(){
             field.focus();
             return false;
-        }).keydown(onkeydown);
+        }).keydown(fieldActions.onkeydown);
 
         // Focus the field as soon as the app is on
         field.focus();
     });
 });
-
-function onkeydown(e){
-    // On Escape
-    if(e.which === 27){
-        clearOrClose();
-    }
-    // On F5
-    else if(e.which === 116){
-        onDirActions.showDir();
-    }
-    // On F6
-    else if(e.which === 117){
-        onDirActions.cmdOnDir();
-    }
-}
-
-function clearOrClose(){
-    if(field.val() === ''){
-        remote.getCurrentWindow().close();
-    }
-    else{
-        clearField();
-    }
-}
-
-function clearField(){
-    field.val('');
-    field.autocomplete("close");
-}
